backend/requirements.txt
# Web Framework
fastapi==0.111.1
uvicorn[standard]==0.29.0
python-multipart==0.0.6
starlette==0.27.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1  # migrations için

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
pyotp==2.9.0
qrcode[pil]==7.4.2

# Email & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0

# Templating
jinja2==3.1.2

# HTTP Client & Async I/O
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1

# Cache & Rate Limiting
redis==5.0.1

# System Monitoring
psutil==5.9.6
GPUtil==1.4.0

# High Performance Computing
asyncio-pool==0.6.0
aiomultiprocess==0.9.0

# Image Processing & Optimization
Pillow==10.1.0
opencv-python==4.8.1.78
imageio==2.31.6

# Enterprise Features
redis==5.0.1
celery==5.3.4

# File Type Detection (Optional)
python-magic==0.4.27

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1  # FastAPI testing

# Type Stubs
types-redis>=4.6.0
types-requests>=2.31.0
sqlalchemy-stubs>=0.4

# Image Processing (QR Code dependencies)
Pillow==10.1.0

# AI and Image Processing
opencv-python==4.8.1.78
numpy==1.24.3
google-cloud-vision==3.4.4  # Google AI Vision API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
